backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
python-multipart
openai
python-dotenv
